module/power_domain: Move notifications into a seperate file
The notifcation functions are now in a new file with a new header.
This is to clean up mod_power_domain.c and get rid of some #ifdefs.
